+++ CHANGELOG.md
@@ -5,11 +5,14 @@ It is based on Tidepool Blip 1.27.
 ## 3.0.5-beta - 2023-07-12
 ### Added
 - YLP-2484 Remove level column in device change parameter history
+
+### Changed
+- YLP-2227 Adjustment of the stats on carbs of the dashboard
+
 ### Fixed
 - YLP-2438 Inconsistency in the daily view between carbs displayed on the graph and Total carbs in the statistics panel
 
 ### Engineering use
-- YLP-2227 Adjustment of the stats on carbs of the dashboard
 - YLP-2573 Create a dedicated message for the caregiver when the account creation is complete
 - YLP-2594 Handle release branches in GitHub CI workflow
 - YLP-2595 Update deprecated GitHub Actions to Node 16